Move some code to _test
