+++ parser/parser.go
@@ -5,7 +5,6 @@ package parser
 import (
 	"cloud.google.com/go/bigquery"
 
-	"github.com/m-lab/etl/bq"
 	"github.com/m-lab/etl/etl"
 	"github.com/m-lab/etl/metrics"
 )
@@ -58,45 +57,5 @@ func (np *NullParser) TableName() string {
 	return "null-table"
 }
 func (np *NullParser) TaskError() error {
-	return nil
-}
-
-//------------------------------------------------------------------------------------
-// TestParser ignores the content, returns a MapSaver containing meta data and
-// "testname":"..."
-// TODO add tests
-type TestParser struct {
-	inserter     etl.Inserter
-	etl.RowStats // Allows RowStats to be implemented through an embedded struct.
-}
-
-func NewTestParser(ins etl.Inserter) etl.Parser {
-	return &TestParser{
-		ins,
-		&FakeRowStats{}} // Use a FakeRowStats to provide the RowStats functions.
-}
-
-func (tp *TestParser) ParseAndInsert(meta map[string]bigquery.Value, testName string, test []byte) error {
-	metrics.TestCount.WithLabelValues("table", "test", "ok").Inc()
-	values := make(map[string]bigquery.Value, len(meta)+1)
-	// TODO is there a better way to do this?
-	for k, v := range meta {
-		values[k] = v
-	}
-	values["testname"] = testName
-	return tp.inserter.InsertRow(bq.MapSaver{Values: values})
-}
-
-// These functions are also required to complete the etl.Parser interface.
-func (tp *TestParser) Flush() error {
-	return nil
-}
-func (tp *TestParser) TableName() string {
-	return "test-table"
-}
-func (tp *TestParser) FullTableName() string {
-	return "test-table"
-}
-func (tp *TestParser) TaskError() error {
 	return nil
 }
